Locales/Keys: Add translation progress bar - addresses #122
Introduce progress for locales.

+++ app/models/Stat.java
@@ -12,9 +12,8 @@ import java.util.UUID;
  */
 @Entity
 @Sql
-public class Stat
-{
+public class Stat {
 	public UUID id;
 
-	public Long count;
+	public Double count;
 }
